[INTERNAL] Refactoring and code cleansing.
