--- include/seqan/align_parallel_2.h
@@ -1,166 +1,0 @@
-// ==========================================================================
-//                 SeqAn - The Library for Sequence Analysis
-// ==========================================================================
-// Copyright (c) 2006-2016, Knut Reinert, FU Berlin
-// All rights reserved.
-//
-// Redistribution and use in source and binary forms, with or without
-// modification, are permitted provided that the following conditions are met:
-//
-//     * Redistributions of source code must retain the above copyright
-//       notice, this list of conditions and the following disclaimer.
-//     * Redistributions in binary form must reproduce the above copyright
-//       notice, this list of conditions and the following disclaimer in the
-//       documentation and/or other materials provided with the distribution.
-//     * Neither the name of Knut Reinert or the FU Berlin nor the names of
-//       its contributors may be used to endorse or promote products derived
-//       from this software without specific prior written permission.
-//
-// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
-// AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
-// IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
-// ARE DISCLAIMED. IN NO EVENT SHALL KNUT REINERT OR THE FU BERLIN BE LIABLE
-// FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
-// DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
-// SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
-// CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
-// LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY
-// OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH
-// DAMAGE.
-//
-// ==========================================================================
-// Author: Rene Rahn <rene.rahn@fu-berlin.de>
-// ==========================================================================
-
-#ifndef INCLUDE_SEQAN_ALIGN_PARALLEL_2_H_
-#define INCLUDE_SEQAN_ALIGN_PARALLEL_2_H_
-
-// ============================================================================
-// STD Prerequisites
-// ============================================================================
-
-#if SEQAN_DEBUG_ENABLED
-#include <typeinfo>
-#include <cxxabi.h>
-#include <stdlib.h>
-#endif
-
-#include <type_traits>
-#include <utility>
-#include <vector>
-#include <unordered_map>
-//#include <thread>
-//#include <mutex>
-#include <shared_mutex>
-//#include <atomic>
-//#include <condition_variable>
-
-#include <seqan/parallel.h>
-
-#ifdef DP_PARALLEL_SHOW_PROGRESS
-
-namespace impl{
-namespace dp_parallel_progress {
-
-std::atomic<unsigned> counter;
-
-void show_progress(unsigned total)
-{
-    using namespace std::chrono_literals;
-
-    auto check_progress = [total]()
-    {
-        unsigned mark = 0;
-        unsigned tick = (total + 9)/10;
-
-        while (counter.load() < total)
-        {
-            unsigned c = counter.load();
-            if (c >= mark * tick)
-            {
-                std::cout << mark * 10 << "% ";
-                mark = (c/tick) + 1;
-            }
-            //std::this_thread::sleep_for(2s);
-        }
-        std::cout << "100%\n";
-    };
-    counter.store(0);
-    std::thread(check_progress).detach();
-}
-}
-}
-#endif  // DP_PARALLEL_SHOW_PROGRESS
-
-// ============================================================================
-// SeqAn Prerequisites
-// ============================================================================
-
-#include <seqan/basic.h>
-#include <seqan/simd.h>
-#include <seqan/align.h>
-
-#include <seqan/align_parallel/dp_parallel_execution_policies.h>
-
-#include <seqan/align/alignment_configurator.h>                 // TODO(rrahn): rename! refactor! Move to align_basic
-#include <seqan/align_parallel/align_parallel_impl.h>                 // TODO(rrahn): rename! refactor! Move to align_basic
-#include <seqan/align_parallel/dp_base_config.h>                // TODO(rrahn): rename! refactor!
-#include <seqan/align_parallel/dp_parallel_base.h>                // TODO(rrahn): rename! refactor!
-
-// ============================================================================
-// Parallel DataStructures
-// ============================================================================
-
-#include <seqan/align_parallel/enumerable_thread_local.h>
-#include <seqan/align_parallel/enumerable_thread_local_limit.h>
-#include <seqan/align_parallel/enumerable_thread_local_iterator.h>
-#include <seqan/align_parallel/dp_parallel_execution_policies.h>      // TODO(rrahn): rename! refactor!
-#include <seqan/align_parallel/parallel_thread_pool_std.h>      // TODO(rrahn): rename! refactor
-#include <seqan/align_parallel/wavefront_simd_dp_tasks.h>  // TODO(rrahn): Write tests!
-
-// ============================================================================
-// Wavefront DataStructures
-// ============================================================================
-
-#include <seqan/align_parallel/wavefront_task_scheduler.h>
-#include <seqan/align_parallel/wavefront_alignment_scheduler.h>
-#include <seqan/align_parallel/wavefront_intermediate_dp_result.h>     // TODO(rrahn): rename! refactor!
-#include <seqan/align_parallel/wavefront_alignment_thread_local_storage.h>
-
-//#include <seqan/align_parallel/parallel_std_task_context.h>     // TODO(rrahn): rename! refactor!
-//#include <seqan/align_parallel/parallel_alignment_instance.h>
-//#include <seqan/align_parallel/parallel_alignment_scheduler.h>
-//#include <seqan/align_parallel/enumerable_thread_specific.h>
-
-// ============================================================================
-// Wavefront Alignment Task
-// ============================================================================
-
-#include <seqan/align_parallel/dp_parallel_scout.h>
-#ifdef SEQAN_SIMD_ENABLED
-#include <seqan/align_parallel/dp_parallel_scout_simd.h>
-#endif
-#include <seqan/align_parallel/wavefront_alignment_task_event.h>
-
-#include <seqan/align_parallel/dp_task_base_2_impl.h>
-#include <seqan/align_parallel/dp_task_base_2.h>
-#include <seqan/align_parallel/wavefront_task_executor.h>
-#include <seqan/align_parallel/wavefront_executor_std.h>
-#include <seqan/align_parallel/wavefront_alignment_task.h>
-
-// ============================================================================
-// DP Task
-// ============================================================================
-
-
-//#include <seqan/align_parallel/dp_parallel_base.h>
-
-// ============================================================================
-// Interfaces
-// ============================================================================
-
-#include <seqan/align_parallel/alignment_interface_impl_wavefront.h>
-#include <seqan/align_parallel/global_alignment_interface.h>
-//#include <seqan/align_parallel/align_parallel_impl.h>
-
-#endif  // #ifndef INCLUDE_SEQAN_ALIGN_PARALLEL_2_H_
