+++ tests/align_parallel/test_align_parallel_interface.h
@@ -34,7 +34,7 @@
 
 // #define DP_PARALLEL_SHOW_PROGRESS // Enable when debugging.
 
-#include <seqan/align_parallel_2.h>
+#include <seqan/align_parallel.h>
 
 #include "../align/test_mock.h"
